+++ src/commands/doctor.ts
@@ -12,8 +12,7 @@ import { Flags, loglevel, SfCommand } from '@salesforce/sf-plugins-core';
 import { Lifecycle, Messages, SfError } from '@salesforce/core';
 import * as open from 'open';
 import got from 'got';
-import * as ProxyAgent from 'proxy-agent';
-import { getProxyForUrl } from 'proxy-from-env';
+import { ProxyAgent } from 'proxy-agent';
 import { Doctor as SFDoctor, SfDoctor, SfDoctorDiagnosis } from '../doctor';
 import { DiagnosticStatus } from '../diagnostics';
 
@@ -63,10 +62,9 @@ export default class Doctor extends SfCommand<SfDoctorDiagnosis> {
 
     this.outputDir = path.resolve(flags['output-dir'] ?? process.cwd());
 
-    // eslint-disable-next-line @typescript-eslint/require-await
     lifecycle.on<DiagnosticStatus>('Doctor:diagnostic', async (data) => {
       this.log(`${data.status} - ${data.testName}`);
-      this.doctor.addDiagnosticStatus(data);
+      return Promise.resolve(this.doctor.addDiagnosticStatus(data));
     });
 
     if (flags.command) {
@@ -120,7 +118,7 @@ export default class Doctor extends SfCommand<SfDoctorDiagnosis> {
       const raw = 'https://raw.githubusercontent.com/forcedotcom/cli/main/.github/ISSUE_TEMPLATE/bug_report.md';
       const ghIssue = await got(raw, {
         throwHttpErrors: false,
-        agent: { https: ProxyAgent(getProxyForUrl(raw)) },
+        agent: { https: new ProxyAgent() },
       });
 
       const title = (