+++ src/commands/info/releasenotes/display.ts
@@ -5,9 +5,6 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-// Needed this to ensure the "helpers" were declared before read in examples
-/* eslint-disable @typescript-eslint/member-ordering */
-
 import * as os from 'os';
 import { marked } from 'marked';
 import * as TerminalRenderer from 'marked-terminal';
@@ -23,22 +20,19 @@ import { parseReleaseNotes } from '../../../shared/parseReleaseNotes';
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
 
-const HIDE_NOTES = 'SFDX_HIDE_RELEASE_NOTES';
-const HIDE_FOOTER = 'SFDX_HIDE_RELEASE_NOTES_FOOTER';
+const helpers = ['stable', 'stable-rc', 'latest', 'latest-rc', 'rc'];
 
 // Load the specific messages for this file. Messages from @salesforce/command, @salesforce/core,
 // or any library that is using the messages framework can also be loaded this way.
 const messages = Messages.loadMessages('@salesforce/plugin-info', 'display');
 
 export default class Display extends SfCommand<DisplayOutput | undefined> {
-  private static helpers = ['stable', 'stable-rc', 'latest', 'latest-rc', 'rc'];
-
   public static readonly summary = messages.getMessage('summary');
   public static readonly description = messages.getMessage('description');
 
   public static readonly aliases = ['whatsnew'];
 
-  public static readonly examples = messages.getMessages('examples', [Display.helpers.join(', ')]);
+  public static readonly examples = messages.getMessages('examples', [helpers.join(', ')]);
 
   public static readonly flags = {
     version: Flags.string({
@@ -53,13 +47,14 @@ export default class Display extends SfCommand<DisplayOutput | undefined> {
   };
 
   public async run(): Promise<DisplayOutput | undefined> {
+    const HIDE_NOTES = this.config.bin === 'sf' ? 'SF_HIDE_RELEASE_NOTES' : 'SFDX_HIDE_RELEASE_NOTES';
+    const HIDE_FOOTER = this.config.bin === 'sf' ? 'SF_HIDE_RELEASE_NOTES_FOOTER' : 'SFDX_HIDE_RELEASE_NOTES_FOOTER';
+
     const logger = Logger.childFromRoot(this.constructor.name);
     const { flags } = await this.parse(Display);
     const env = new Env();
 
-    const isHook = flags.hook;
-
-    if (env.getBoolean(HIDE_NOTES) && isHook) {
+    if (env.getBoolean(HIDE_NOTES) && flags.hook) {
       // We don't ever want to exit the process for info:releasenotes:display (whatsnew)
       // In most cases we will log a message, but here we only trace log in case someone using stdout of the update command
       logger.trace(`release notes disabled via env var: ${HIDE_NOTES}`);
@@ -78,7 +73,7 @@ export default class Display extends SfCommand<DisplayOutput | undefined> {
 
       let version = flags.version ?? installedVersion;
 
-      if (Display.helpers.includes(version)) {
+      if (helpers.includes(version)) {
         version = await getDistTagVersion(distTagUrl, version);
       }
 
@@ -100,7 +95,7 @@ export default class Display extends SfCommand<DisplayOutput | undefined> {
         this.log(marked.parser(tokens));
       }
 
-      if (isHook) {
+      if (flags.hook) {
         if (env.getBoolean(HIDE_FOOTER)) {
           await Lifecycle.getInstance().emitTelemetry({ eventName: 'FOOTER_HIDDEN' });
         } else {
@@ -109,7 +104,7 @@ export default class Display extends SfCommand<DisplayOutput | undefined> {
         }
       }
     } catch (err) {
-      if (isHook) {
+      if (flags.hook) {
         // Do not throw error if --hook is passed, just warn so we don't exit any processes.
         // --hook is passed in the post install/update scripts
         const { message, stack, name } = err as Error;